backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
supabase==2.5.0
python-multipart==0.0.19
pydantic==2.10.5
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
mangum==0.18.0
psycopg2-binary==2.9.11

